Address code review: extract OrderDetail, move shiftWeek, comment action
- Extract OrderDetail subcomponent to src/components/admin/order-detail.tsx
  (order-row.tsx: 274 → 171 lines, under CLAUDE.md 200-line cap).
- Move shiftWeek into src/lib/week.ts so all week math sits in one file.
- Comment advanceOrderStatus naming the admin_all_orders RLS policy it
  relies on, so a future reader doesn't "fix" the cookie client to admin.
